Nested the route link inside of a ternary operator
