+++ src/App.js
@@ -1,5 +1,3 @@
-import axios from 'axios';
-import { useEffect, useState } from 'react';
 import UserSelectionForm from './components/UserSelectionForm.js';
 import { Route, Routes } from 'react-router-dom';
 import PlayerNames from './components/PlayerNames.js';
@@ -8,98 +6,15 @@ import PlayerNames from './components/PlayerNames.js';
 import './styles/sass/App.scss';
 
 function App() {
-  // api call for category
-  const [categoryArr, setCategoryArr] = useState([]);
-
-  // useStates from form component - for second api call
-  const [numOfPlayers, setNumOfPlayers] = useState(0);
-
-  const [userCategory, setUserCategory] = useState("");
-  const [userDifficulty, setUserDifficulty] = useState("");
-
-  const [submitButton, setSubmitButton] = useState(false)
-
-  const handlePlayerNumber = (event) => {
-    setNumOfPlayers(event.target.value);
-  }
-
-  const handleCategoryChoice = (event) => {
-    setUserCategory(event.target.value);
-    // console.log(userCategory);
-  }
-
-  const handleDifficultyChoice = (event) => {
-    setUserDifficulty(event.target.value);
-    // console.log(userDifficulty);
-  }
-
-  const submitHandler = (event) => {
-      event.preventDefault();
-      setSubmitButton(!submitButton);
-    };
-
-  // useEffect for axios - put this info in dropdown
-  // associate id and name of category
-  useEffect(() => {
-    axios({
-      url: "https://opentdb.com/api_category.php",
-      method: "GET",
-      responseType: "json",
-      params: {},
-    })
-      .then((res) => {
-        setCategoryArr(res.data.trivia_categories);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  }, [])
-
-
-useEffect( () => {
-  if (userCategory !== ""){
-    
-    axios({
-        url: "https://opentdb.com/api.php",
-        method: "GET",
-        responseType: "json",
-        params: {
-          // category: 19,
-          // type: "multiple",
-          // difficulty: "hard",
-          amount: 10,
-          category: userCategory,
-          type: "multiple",
-          difficulty: userDifficulty,
-        },
-      })
-      .then((res) => {
-        console.log(res.data.results);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  }
-}, [submitButton])
 
   return (
     <div>
       <h1>RoboTrivia</h1>
-      <UserSelectionForm 
-        array={categoryArr} 
-        submitHandler={submitHandler} 
-        handlePlayerNumber={handlePlayerNumber}
-        handleCategoryChoice={handleCategoryChoice} 
-        handleDifficultyChoice={handleDifficultyChoice}
-        numOfPlayers={numOfPlayers}
-        userCategory={userCategory}
-        userDifficulty={userDifficulty} 
-      />
 
       <Routes>
-        <Route path='/playernames/:num' element={<PlayerNames />} />
+        <Route path="/" element={<UserSelectionForm />} />
+        <Route path="/playernames/:num" element={<PlayerNames />} />
       </Routes>
-
     </div>
   );
 }
